# [nnx] optimize Variable