… docs_nnx/nnx_basics.ipynb
… tests/nnx/containers_test.py
… tests/nnx/module_test.py
… tests/nnx/spmd_test.py
… uv.lock