# What does this PR do?
* Optimizes the `Variable` and `VariableState` types. In particular, instead of adding the metadata as `Variable` / `VariableState` attributes, each now has a `_var_metadata` field that stores the metadata. `__getattr__` and `__setattr__` are modified to retrieve the values at their new location.
* Removes `*_hooks` attributes and replaces with a simpler, single-hook metadata with the form `on_*` e.g. `on_add_axis`, `on_remove_axis`, etc.
```
python benchmarks/nnx_graph_overhead.py --mode=all --depth=100 --total_steps=1000

### NNX ###
total time: 6.162761926651001
time per step: 6162.76 µs
time per layer: 61.63 µs

### JAX ###
total time: 1.4132189750671387
time per step: 1413.22 µs
time per layer: 14.13 µs
```